Parsing: added support for binary expression
Operator now presents precedence(), associativity() and chainability()
APIs. These well be useful in the Parser and also for displaying a
table of operator precedence in the future.

Added the following expressions to the ast: NullExpression,
UndefinedExpression, UnreachableExpression.

Added BinaryExpression to the ast.

Parser can now parse the new ast expressions.